Crate owsql

👎 Deprecated

OverwriteSQL

owsql is a secure library for PostgreSQL, MySQL and SQLite.
Unlike other libraries, you can use string concatenation to prevent SQL injection.

let id_input = "42 OR 1=1; --";
let sql = conn.ow("SELECT name FROM users WHERE id = ") + id_input;
println!("[{}]", sql); // [ OWSQL47xyz6km0CfbRt0BA38Z2DxrleESyPPg4 42 OR 1=1; --]
// At runtime it will be transformed into a query like
// "SELECT name FROM users WHERE id = '42 OR 1=1; --'".

Example

Open a connection of SQLite, create a table, and insert some rows:

let mut conn = owsql::sqlite::open(":memory:").unwrap();
let stmt = conn.ow(r#"CREATE TABLE users (name TEXT, age INTEGER);
              INSERT INTO users (name, age) VALUES ('Alice', 42);
              INSERT INTO users (name, age) VALUES ('Bob', 69);"#);
conn.execute(&stmt).unwrap();

Select some rows and process them one by one as plain text:

let age = "50";
let sql = conn.ow("SELECT * FROM users WHERE age > ") + age;
conn.iterate(&sql, |pairs| {
    for &(column, value) in pairs.iter() {
        println!("{} = {}", column, value.unwrap());
    }
    true
}).unwrap();

It can be executed after getting all the rows of the query:

let age = "50";
let sql = conn.ow("SELECT * FROM users WHERE age > ") + age;
let rows = conn.rows(&sql).unwrap();
for row in rows.iter() {
    println!("name = {}", row.get("name").unwrap_or("NULL"));
}

Modules

constants	Deprecated

mysql	Deprecated
Interface to MySQL of OverwriteSQL.

postgres	Deprecated
Interface to PostgreSQL of OverwriteSQL.

sqlite	Deprecated
Interface to SQLite of OverwriteSQL.

Macros

params	Deprecated
This macro is a convenient way to pass named parameters to a statement.

sanitize_like	Deprecated
Sanitizes a string so that it is safe to use within an SQL LIKE statement.
This method uses escape_character to escape all occurrences of '_' and '%'.

Structs

Connection	Deprecated
A database connection.

Row	Deprecated
A single result row of a query.

Enums

OwsqlError	Deprecated
Enum listing possible errors from owsql.

OwsqlErrorLevel	Deprecated
Change the output error message.

Traits

IntoInner	Deprecated

Functions

html_special_chars	Deprecated
Convert special characters to HTML entities.

Type Definitions

Result	Deprecated
A typedef of the result returned by many methods.